requirements.txt
Flask
Flask_SQLAlchemy
numpy
Pillow
tensorflow
scikit-learn